Iteration would cause growing of the stored working set of defs etc.

After:
```

     Running benches/parser.rs (target/release/deps/parser-aac4614e0ba8b910)
Benchmarking parse config: Warming up for 3.0000 s
Warning: Unable to complete 100 samples in 5.0s. You may wish to increase target time to 9.3s, enable flat sampling, or reduce sample count to 50.
parse config            time:   [1.8112 ms 1.8139 ms 1.8163 ms]
                        change: [-0.6209% -0.3693% -0.1163%] (p = 0.00 < 0.05)
                        Change within noise threshold.
Found 2 outliers among 100 measurements (2.00%)
  2 (2.00%) high severe
```
Before:
```
     Running benches/parser.rs (target/release/deps/parser-aac4614e0ba8b910)
parse config            time:   [116.71 ms 122.65 ms 128.65 ms]
                        change: [+6322.4% +6676.1% +7012.2%] (p = 0.00 < 0.05)
                        Performance has regressed.
```